backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
supabase>=2.3.0
openai>=1.12.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.116.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
httpx>=0.24.0
python-multipart>=0.0.6
